Add Weather projects and refactor other projects
Added two new projects `Weather` and `Weather2` that use adapters to retrieve WeekForecast from different weather providers and output the result to different destinations. Also, `shared` library is added which have the method used to retrieve HTTP request in `JSON to XML Converter` project. All the classes of `Duck Simulator` are refactored and sorted into different packages named `adapter`, `models`, and `interfaces`. Modified `encoder.xml`, `compiler.xml` and `modules.xml` to reflect the changes made in the projects. Grain size is reduced to improve the readability, maintenance and scalability of the projects.